fix: split consume same-name mapper test to satisfy backend file limit
